assign auth api to header component
